Add equations 3.1 through 3.7 to chapter 3
There is a symbol in equation 3.1 akin to rho that I could not figure out how to add, and it is currently represented with a "P". I was also unable to figure out how to format the indentations of the equations as in the original book, and have organized the format as I could.